Decouple WAL delta resolve method into standalone function
